+++ packages/plugins/lib/helpers/router/index.js
@@ -1,19 +1,24 @@
 import { jsx as _jsx, Fragment as _Fragment } from "react/jsx-runtime";
-import { lazy } from "react";
+import { lazy, useEffect } from "react";
 import { Route, Routes, Outlet, useMatch, useNavigate } from "react-router-dom";
-export const DynamicRouteProvider = ({ routes, }) => {
+export const RouterItem = ({ route, render, key, }) => {
     const navigate = useNavigate();
+    const match = useMatch(route.path || "");
+    useEffect(() => {
+        if (route.path && match) {
+            document.title = route.title || "";
+            if (route.redirect && route.path !== route.redirect) {
+                navigate(route.redirect, { replace: true });
+            }
+        }
+    }, []);
+    const LazyComponent = lazy(() => import(`@/${route.component}`));
+    return (_jsx(Route, { path: route.path || "*", element: _jsx(_Fragment, { children: route.routes ? (_jsx(LazyComponent, { ...route, children: _jsx(Outlet, {}) })) : (_jsx(LazyComponent, {})) }), children: route.routes ? render(route.routes) : undefined }, key));
+};
+export const DynamicRouteProvider = ({ routes, }) => {
     const render = (list) => {
         return list.map((ele, i) => {
-            const match = useMatch(ele.path || "");
-            if (ele.path && match) {
-                document.title = ele.title || "";
-                if (ele.redirect && ele.path !== ele.redirect) {
-                    navigate(ele.redirect, { replace: true });
-                }
-            }
-            const LazyComponent = lazy(() => import(`@/${ele.component}`));
-            return (_jsx(Route, { path: ele.path || "*", element: _jsx(_Fragment, { children: ele.routes ? (_jsx(LazyComponent, { ...ele, children: _jsx(Outlet, {}) })) : (_jsx(LazyComponent, {})) }), children: ele.routes ? render(ele.routes) : undefined }, ele.name || "key-" + i));
+            return RouterItem({ key: ele.name || "key-" + i, route: ele, render });
         });
     };
     return _jsx(Routes, { children: render(routes) });